+++ CHANGELOG.md
@@ -5,7 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [STVP](https://mia.kiwi/projects/stvp).
 
-## [Unreleased]
+## [25.2.1] - 2025-11-28
+
+### Added
+
+- Button icon location parameter (left or right)
+- Static method `getInputAsHTMLElement` to `Component`
+- Icon buttons
+- Breadcrumbs
+
+### Changed
+
+- Renamed `AccordionsContainer` to `AccordionGroup`
+
+
+
+## [25.2.0] - 2025-11-28
 
 ### Added